requirements.txt
flask
flask-sqlalchemy
flask-script
flask-migrate
flask-login
simple-crypt
